pyproject.toml
[project]
name = "visionpi"
version = "0.1.0"
description = "Person-following rover: Raspberry Pi vision + Arduino Nano R4 motor control"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
# Install on Pi inside venv_hailo_apps (hailo + GStreamer already present there)
vision = [
    "numpy>=1.24",
    "opencv-python>=4.8",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
# Full Pi runtime: pip install -e ".[vision,dev]"
pi = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "pyserial>=3.5",
]

[project.scripts]
follow-person = "apps.follow_person:main"
follow-controller = "apps.follow_controller_cli:main"
motor-test = "scripts.motor_test_rc:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "motors*", "robot*", "vision*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]